Struct x25519_dalek::EphemeralSecret

pub struct EphemeralSecret(_);

A short-lived Diffie-Hellman secret key that can only be used to compute a single SharedSecret.

This type is identical to the StaticSecret type, except that the EphemeralSecret::diffie_hellman method consumes and then wipes the secret key, and there are no serialization methods defined. This means that EphemeralSecrets can only be generated from fresh randomness by EphemeralSecret::new and the compiler statically checks that the resulting secret is used at most once.

Implementations

impl EphemeralSecret

pub fn diffie_hellman(self, their_public: &PublicKey) -> SharedSecret

Perform a Diffie-Hellman key agreement between self and their_public key to produce a SharedSecret.

pub fn new<T: RngCore + CryptoRng>(csprng: T) -> Self

Generate an x25519 EphemeralSecret key.