Struct x25519_dalek::StaticSecret

pub struct StaticSecret(/* private fields */);

A Diffie-Hellman secret key that can be used to compute multiple SharedSecrets.

This type is identical to the EphemeralSecret type, except that the StaticSecret::diffie_hellman method does not consume the secret key, and the type provides serialization methods to save and load key material. This means that the secret may be used multiple times (but does not have to be).

Warning

If youโre uncertain about whether you should use this, then you likely should not be using this. Our strongly recommended advice is to use EphemeralSecret at all times, as that type enforces at compile-time that secret keys are never reused, which can have very serious security implications for many protocols.

Implementations

impl StaticSecret

pub fn diffie_hellman(&self, their_public: &PublicKey) -> SharedSecret

Perform a Diffie-Hellman key agreement between self and their_public key to produce a SharedSecret.

pub fn new<T: RngCore + CryptoRng>(csprng: T) -> Self

๐Deprecated since 2.0.0: Renamed to random_from_rng. This will be removed in 2.1.0

Generate a new StaticSecret with the supplied RNG.

pub fn random_from_rng<T: RngCore + CryptoRng>(csprng: T) -> Self

Generate a new StaticSecret with the supplied RNG.

pub fn to_bytes(&self) -> [u8; 32]

Extract this keyโs bytes for serialization.

pub fn as_bytes(&self) -> &[u8; 32]

View this key as a byte array.