Crate x25519_dalek

Note that docs will only build on nightly Rust until feature(external_doc) is stabilized.

Structs

EphemeralSecret
A EphemeralSecret is a short lived Diffie-Hellman secret key used to create a SharedSecret when given their PublicKey.

PublicKey
A PublicKey is the corresponding public key converted from an EphemeralSecret or a StaticSecret key.

SharedSecret
A SharedSecret is a Diffie-Hellman shared secret that’s generated from your EphemeralSecret or StaticSecret and their PublicKey.

StaticSecret
A StaticSecret is a static Diffie-Hellman secret key that can be saved and loaded to create a SharedSecret when given their PublicKey.

Constants

X25519_BASEPOINT_BYTES
The X25519 basepoint, for use with the bare, byte-oriented x25519 function. This is provided for people who cannot use the typed DH API for some reason.

Functions

x25519
The bare, byte-oriented x25519 function, exactly as specified in RFC7748.