Crate x25519_dalek

Note that docs will only build on nightly Rust until feature(external_doc) is stabilized.

Structs

EphemeralSecret

A short-lived Diffie-Hellman secret key that can only be used to compute a single SharedSecret.

PublicKey

A Diffie-Hellman public key, corresponding to an EphemeralSecret or StaticSecret key.

SharedSecret

The result of a Diffie-Hellman key exchange.

StaticSecret

A Diffie-Hellman secret key that can be used to compute multiple SharedSecrets.

Constants

X25519_BASEPOINT_BYTES

The X25519 basepoint, for use with the bare, byte-oriented x25519 function. This is provided for people who cannot use the typed DH API for some reason.

Functions

x25519

The bare, byte-oriented x25519 function, exactly as specified in RFC7748.